@@ -23,28 +23,22 @@
   },
   "dependencies": {},
   "peerDependencies": {
-    "prop-types": "^15.5.4",
-    "react": "^15.0.0 || ^16.0.0",
-    "react-dom": "^15.0.0 || ^16.0.0"
+    "react": "^16.8.0-alpha.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/runtime": "^7.0.0",
-    "@svgr/rollup": "^2.4.1",
+    "@babel/core": "^7.2.2",
+    "@babel/runtime": "^7.3.1",
     "@types/jest": "^23.1.5",
     "@types/react": "^16.3.13",
-    "@types/react-dom": "^16.0.5",
     "cross-env": "^5.1.4",
     "gh-pages": "^1.2.0",
     "react": "^16.4.1",
-    "react-dom": "^16.4.1",
     "react-scripts-ts": "^2.16.0",
     "rollup": "^0.62.0",
     "rollup-plugin-babel": "^4.0.1",
     "rollup-plugin-commonjs": "^9.1.3",
     "rollup-plugin-node-resolve": "^3.3.0",
     "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-postcss": "^1.6.2",
     "rollup-plugin-typescript2": "^0.17.0",
     "rollup-plugin-url": "^1.4.0",
     "typescript": "^2.8.3"
