@@ -14,7 +14,7 @@ module.exports = async () => {
   const defaults = await getDefaultLibraryParams()
 
   program
-    .name('create-react-library')
+    .name('create-react-hook')
     .version(version)
     .usage('[options] [package-name]')
     .option('-d, --desc <string>', 'package description')
