@@ -1,13 +1,13 @@
-import React, { Component } from 'react'
+import React from 'react'
 
-import ExampleComponent from '{{name}}'
+import useMyHook from '{{name}}'
 
-export default class App extends Component {
-  render () {
-    return (
-      <div>
-        <ExampleComponent text='Modern React component module' />
-      </div>
-    )
-  }
+const App = () => {
+  const example = useMyHook()
+  return (
+    <div>
+      {example}
+    </div>
+  )
 }
+export default App
